Fixed an issue where some event ID's were exceeding the maximum value for int32 causing an overflow exception Cleaned up import statements

+++ Entities/Event.cs
@@ -1,14 +1,12 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using EventbriteNET.HttpApi;
 
 namespace EventbriteNET.Entities
 {
     public class Event : EntityBase
     {
-        public int Id;
+        public long Id;
         public string Title;
         public string Description;
         public DateTime StartDateTime;
